Initial commit - Added Python script for MySQL to Google Sheets automation
This commit introduces a Python script that automates the process of fetching data from a MySQL database, transforming it using Pandas, and finally, inserting it into Google Sheets.

Key functionalities include:
- Connection to a MySQL database and retrieval of data based on a provided SQL query
- Data transformation via Pandas to format date and float columns
- Connection to a Google Sheets document using OAuth2 and insertion of the data into the first worksheet

The script is designed for flexibility, allowing customization of the SQL query and the specific Google Sheets document used. Sensitive information such as database credentials and the Google Sheets URL are to be supplied by the user.

Future enhancements can include the addition of notifications upon successful data transfers and scheduling the script for automated data updates.